+++ client/src/pages/Home.js
@@ -1,23 +1,14 @@
 import React from 'react';
-import Login from '../components/Login';
-import Signup from '../components/Signup';
-import Auth from '../utils/auth';
 
-import { Container, Col } from 'react-bootstrap';
-import { useQuery } from '@apollo/client';
-import { QUERY_THOUGHTS, QUERY_ME_BASIC } from '../utils/queries';
+import { Container } from 'react-bootstrap';
 
 const Home = () => {
 
   return (
     <main>
       <Container>
-        <div>
-          <Login />
-        </div>
-        <div>
-          <Signup />
-        </div>
+        <h2 className='purple-bg'>Create your business's own COUPON MAGIC today!</h2><br />
+        <h4 className='purple-bg'>Sign in or create an account to be taken to the coupon creation screen.</h4>
       </Container>
     </main>
   );
